+++ CHANGES.md
@@ -3,6 +3,8 @@ Change Log
 
 #### next release (8.2.5)
 
+
+* Add Google Analytics event for drag and drop of files onto map.
 * Allow users to choose whether Story is included in Share
 * Fixed bug that broke Cesium when WebGL was not available. Reverts to Leaflet.
 * Fixed bug where `new Terria()` constructror would try to access `document` and throw an error when running in NodeJS.